user friendly ui

+++ app.js
@@ -72,15 +72,15 @@ async function listenToPeer(pubKey) {
     if (pubKey === node_keys.pub) return; // don't listen to self
 
     gun.get('~' + pubKey).get('blockchain').on(async (data, key) => {
-        let peerBlockchain = Blockchain.fromJSON(JSON.parse(data));
+		let peerBlockchain = Blockchain.fromJSON(JSON.parse(data));
+        console.log(await peerBlockchain.getTotalWork(), await blockchain.getTotalWork())
         if (!await validateFullChain(peerBlockchain)) return;
-        // if (peerBlockchain.length() <= blockchain.length()) return;
-        if (await peerBlockchain.getTotalWorkAverage() <= await blockchain.getTotalWorkAverage()) return;
+        if (await peerBlockchain.getTotalWork() <= await blockchain.getTotalWork()) return;
         console.log('✅⛓️🤝 Better blockchain received from peer:', pubKey);
         Object.assign(blockchain, peerBlockchain);
         await mempool.cleanUp();
         broadcastMyChain();
-        
+
         resetWorker();
         mine(address);
     });
@@ -130,13 +130,16 @@ async function broadcastMyMempool() {
 
 let worker = new Worker('mining.worker.js');
 function resetWorker() {
-    worker.terminate();
+	if (worker) {
+		worker.terminate();
+    }
     worker = new Worker('mining.worker.js');
 }
 
-let mining = true;
+let mining = false;
 async function mine(miningAddress) {
-    if (!mining) return;
+	if (!mining) return;
+    await mempool.cleanUp();
     worker.postMessage({blockchain: blockchain.toJSON(), rewardAddress: miningAddress, mempool: mempool.serialize()});
     worker.onmessage = async function(e) {
         const block = Block.fromJSON(e.data.block);
@@ -149,12 +152,13 @@ async function mine(miningAddress) {
         console.log('✅📦 Mined new block:', await block.hash());
         console.log('Blockchain target:', blockchain.target);
         console.log('Blockchain bytes size:', roughSizeOfObject(blockchain), 'bytes');
-        console.log('Blockchain total work average:', await blockchain.getTotalWorkAverage());
+        console.log('Blockchain total work average:', await blockchain.getTotalWork());
 
         worker.postMessage({blockchain: blockchain.toJSON(), rewardAddress: miningAddress, mempool: mempool.serialize()});
         broadcastMyChain();
-    }
+	}
 }
+
 function toggleMining() {
     if (mining) {
         mining = false;
@@ -162,23 +166,25 @@ function toggleMining() {
         console.log('⛏️🛑[MINER] Stopped mining.');
     } else {
         mining = true;
-        console.log('⛏️🟢[MINER] Resumed mining.');
+		console.log('⛏️🟢[MINER] Resumed mining.');
+		resetWorker();
         mine(address);
     }
 }
 
 async function updateUI() {
-    let balance = blockchain.getConfirmedBalance(address) / UNITS_PER_COIN + ' PC';
-    document.getElementById('debug-wallet-address').innerText = address;
-    document.getElementById('debug-wallet-balance').innerText = balance;
-    document.getElementById('debug-blockchain').innerHTML = JSON.stringify(Array.from(blockchain.chain).reverse(), null, 2);
-    document.getElementById('debug-mempool').innerHTML = await Promise.all(mempool.transactions.map(async tx => await Transaction.fromJSON(tx).hash() + '<br>' + JSON.stringify(tx))).then(results => results.join('<br><br>'));
+	let balance  = blockchain.getConfirmedBalance(address);
+    document.getElementById('debug-wallet-address').textContent = address;
+    document.getElementById('debug-wallet-balance').textContent = balance;
+    document.getElementById('debug-blockchain').textContent = JSON.stringify(Array.from(blockchain.chain).reverse(), null, 2);
+	document.getElementById('debug-mempool').textContent = await Promise.all(mempool.transactions.map(async tx => await Transaction.fromJSON(tx).hash() + '<br>' + JSON.stringify(tx))).then(results => results.join('<br><br>'));
 
     UI.update(
         address,
         balance,
         blockchain,
-        mempool
+		mempool,
+		mining
     );
 }
 
@@ -188,8 +194,6 @@ async function sendTransaction(recipientAddress, amount) {
     await transaction.sign(keys.priv);
     mempool.addTransaction(transaction);
     await broadcastMyMempool(transaction);
-    resetWorker();
-    mine(address);
     console.log('✅📨 Created and broadcasted transaction:', transaction);
 }
 
@@ -200,7 +204,17 @@ document.getElementById('debug-send-button').addEventListener('click', (e) => {
     sendTransaction(recipientAddress, amount);
 });
 
-mine(address);
+root.querySelector("#main-miner-toggle").addEventListener('click', () => {
+	toggleMining();
+});
+
+root.querySelector("#send-panel-submit").addEventListener('click', () => {
+	let recipientAddress = root.querySelector("#send-recipient-address").value;
+	let amount = parseFloat(root.querySelector("#send-amount").value);
+	sendTransaction(recipientAddress, amount);
+	UI.gotoRootView("main");
+});
+
 setInterval(async () => {
     await updateUI();
-}, 1000);
+}, 1500);
